# backend-api/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/database/type": {
      "get": {
        "operationId": "DatabaseController_findTypeList",
        "summary": "데이터베이스 타입 목록 조회",
        "description": "지원하는 데이터베이스 타입의 목록을 조회합니다.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "데이터베이스 타입 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "value": {
                        "type": "string",
                        "example": "postgresql"
                      },
                      "label": {
                        "type": "string",
                        "example": "PostgreSQL"
                      },
                      "port": {
                        "type": "number",
                        "example": 5432
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/v1/database/data": {
      "get": {
        "operationId": "DatabaseController_findData",
        "summary": "데이터 조회",
        "description": "데이터셋 타입에 따라 데이터를 조회합니다.",
        "parameters": [
          {
            "name": "datasetType",
            "required": true,
            "in": "query",
            "description": "데이터셋 타입",
            "schema": {
              "enum": [
                "DATASET",
                "TABLE"
              ],
              "type": "string"
            }
          },
          {
            "name": "databaseId",
            "required": true,
            "in": "query",
            "description": "데이터베이스 ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "datasetId",
            "required": false,
            "in": "query",
            "description": "데이터셋 ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "tableName",
            "required": false,
            "in": "query",
            "description": "테이블 이름",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "데이터 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/v1/database/info/{id}": {
      "get": {
        "operationId": "DatabaseController_findOneInfo",
        "summary": "데이터베이스 연결정보 조회",
        "description": "데이터베이스의 기본 연결 정보를 조회합니다.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "데이터베이스 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "데이터베이스 연결정보"
          },
          "404": {
            "description": "데이터베이스를 찾을 수 없습니다."
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/v1/database": {
      "post": {
        "operationId": "DatabaseController_create",
        "summary": "데이터베이스 연결 생성",
        "description": "새로운 데이터베이스 연결을 생성합니다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "데이터베이스 연결이 성공적으로 생성되었습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDatabaseDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      },
      "get": {
        "operationId": "DatabaseController_findAll",
        "summary": "데이터베이스 목록 조회",
        "description": "등록된 모든 데이터베이스 연결의 목록을 조회합니다.",
        "parameters": [
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "반환할 필드 선택",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "데이터베이스 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "engine": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/v1/database/test": {
      "post": {
        "operationId": "DatabaseController_testConnection",
        "summary": "데이터베이스 연결 테스트",
        "description": "데이터베이스 연결 설정이 올바른지 테스트합니다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "연결 테스트 결과",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Connection successful"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "연결 실패"
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/v1/database/execute": {
      "post": {
        "operationId": "DatabaseController_executeQuery",
        "summary": "SQL 쿼리 실행",
        "description": "지정된 데이터베이스에서 SQL 쿼리를 실행합니다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryExecuteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "쿼리 실행 결과",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 SQL 쿼리"
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/v1/database/{id}": {
      "get": {
        "operationId": "DatabaseController_findOne",
        "summary": "데이터베이스 상세 조회",
        "description": "데이터베이스의 상세 정보와 테이블, 데이터셋을 조회합니다.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "데이터베이스 ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "반환할 필드 선택",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "데이터베이스 상세 정보"
          },
          "404": {
            "description": "데이터베이스를 찾을 수 없습니다."
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      },
      "put": {
        "operationId": "DatabaseController_update",
        "summary": "데이터베이스 연결 수정",
        "description": "데이터베이스 연결 정보를 수정합니다.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "데이터베이스 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatabaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "데이터베이스 연결이 성공적으로 수정되었습니다."
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "404": {
            "description": "데이터베이스를 찾을 수 없습니다."
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      },
      "delete": {
        "operationId": "DatabaseController_remove",
        "summary": "데이터베이스 연결 삭제",
        "description": "데이터베이스 연결을 삭제합니다.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "데이터베이스 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "데이터베이스 연결이 성공적으로 삭제되었습니다."
          },
          "404": {
            "description": "데이터베이스를 찾을 수 없습니다."
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/v1/database/execute/stream": {
      "post": {
        "operationId": "DatabaseController_executeStreamingQuery",
        "summary": "스트리밍 SQL 쿼리 실행",
        "description": "대용량 데이터를 위한 스트리밍 방식으로 SQL 쿼리를 실행합니다.",
        "parameters": [
          {
            "name": "Transfer-Encoding",
            "in": "header",
            "description": "chunked",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryExecuteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "스트리밍 데이터",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "tags": [
          "데이터베이스"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "VanillaMeta API Documentation",
    "description": "VanillaMeta는 기업용 비즈니스 인텔리전스(BI) 웹 애플리케이션으로, 사용자가 코드 작성 없이 다양한 데이터베이스에 연결하여 시각화를 생성하고 대시보드를 구축할 수 있습니다.\n\n## 주요 기능\n- 다중 데이터베이스 연결 지원\n- 50개 이상의 차트 타입 제공\n- 대시보드 생성 및 공유\n- 실시간 데이터 시각화\n\n## 인증\nJWT Bearer 토큰을 사용합니다. 로그인 API를 통해 액세스 토큰을 발급받은 후 Authorization 헤더에 Bearer {token} 형식으로 전달하세요.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/v1",
      "description": "로컬 개발 서버"
    },
    {
      "url": "https://dev-api.vanillameta.com/v1",
      "description": "개발 서버"
    },
    {
      "url": "https://api.vanillameta.com/v1",
      "description": "프로덕션 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT 액세스 토큰을 입력하세요",
        "in": "header"
      }
    },
    "schemas": {
      "CreateDatabaseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "데이터베이스 연결 이름",
            "example": "프로덕션 MySQL DB",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "데이터베이스 설명",
            "example": "메인 서비스용 프로덕션 데이터베이스"
          },
          "connectionConfig": {
            "type": "string",
            "description": "데이터베이스 연결 설정 (JSON 형식)",
            "example": "{\"host\":\"localhost\",\"port\":3306,\"username\":\"root\",\"password\":\"password\",\"database\":\"mydb\"}",
            "format": "json"
          },
          "engine": {
            "type": "string",
            "description": "데이터베이스 엔진",
            "example": "mysql",
            "enum": [
              "mysql",
              "postgresql",
              "mariadb",
              "oracle",
              "mssql",
              "sqlite",
              "bigquery",
              "snowflake",
              "redshift"
            ]
          },
          "type": {
            "type": "string",
            "description": "데이터베이스 타입",
            "example": "mysql",
            "enum": [
              "mysql",
              "postgresql",
              "mariadb",
              "oracle",
              "mssql",
              "sqlite",
              "bigquery",
              "snowflake",
              "redshift"
            ]
          },
          "timezone": {
            "type": "string",
            "description": "타임존 설정",
            "example": "Asia/Seoul",
            "default": "Asia/Seoul"
          }
        },
        "required": [
          "name",
          "connectionConfig"
        ]
      },
      "QueryParameterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "user_id",
            "description": "매개변수 이름"
          },
          "value": {
            "type": "string",
            "example": "123",
            "description": "매개변수 값"
          },
          "type": {
            "type": "string",
            "example": "number",
            "description": "매개변수 타입 (string, number, date)"
          }
        },
        "required": [
          "name",
          "value"
        ]
      },
      "QueryExecuteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "database id"
          },
          "query": {
            "type": "string",
            "example": "SELECT * FROM sample_table1 WHERE id = ?",
            "description": "실행할 SQL 쿼리 (SELECT문만 허용)",
            "maxLength": 10000
          },
          "parameters": {
            "description": "쿼리 매개변수 (prepared statement용)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryParameterDto"
            }
          },
          "limit": {
            "type": "number",
            "example": 1000,
            "description": "최대 결과 행 수 (기본값: 1000)",
            "default": 1000
          }
        },
        "required": [
          "id",
          "query"
        ]
      },
      "UpdateDatabaseDto": {
        "type": "object",
        "properties": {}
      },
...
    },
...
  },
  "security": [
    {
      "AccessToken": []
    }
  ]
}